+++ util/src/lib.rs
@@ -7,6 +7,11 @@
 //!
 //! ## Features
 //!
+//! ### `link`
+//!
+//! Provides implementations for parsing and formatting entities' URLs, such as
+//! webhook URLs.
+//!
 //! ### `permission-calculator`
 //!
 //! Allows the use of a calculator to determine the permissions of a member in